requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
exa-py==1.0.7
cerebras-cloud-sdk==1.1.0
geopy==2.4.0
python-multipart==0.0.6
aiofiles==23.2.1